split useRecentLocationUsers into it's own file
