+++ src/hooks/users/index.ts
@@ -1,12 +1,7 @@
-import { useMemo } from "react";
-
-import { normalizeTimestampToMilliseconds } from "utils/time";
-
 import { useConnectCurrentVenueNG } from "hooks/useConnectCurrentVenueNG";
-import { useUserLastSeenThreshold } from "hooks/useUserLastSeenThreshold";
 import { useVenueId } from "hooks/useVenueId";
 
-import { useWorldUsers } from "./useWorldUsers";
+import { useRecentLocationUsers } from "./useRecentLocationUsers";
 
 export { useConnectWorldUsers } from "./useConnectWorldUsers";
 export { useWorldUsers } from "./useWorldUsers";
@@ -15,34 +10,7 @@ export {
   useWorldUsersByIdWorkaround,
 } from "./useWorldUsersById";
 export { useRecentWorldUsers } from "./useRecentWorldUsers";
-
-/**
- * @description this hook's filtering world users based on their @lastSeenIn location
- *
- * @param locationName is a key for `lastSeenIn` firestore field in user's object
- *
- * @example useRecentLocationUsers(venue.name)
- * @example useRecentLocationUsers(`${venue.name}/${roomTitle}`)
- */
-export const useRecentLocationUsers = (locationName?: string) => {
-  const lastSeenThreshold = useUserLastSeenThreshold();
-  const { worldUsers, isWorldUsersLoaded } = useWorldUsers();
-
-  return useMemo(
-    () => ({
-      recentLocationUsers: locationName
-        ? worldUsers.filter(
-            (user) =>
-              user.lastSeenIn?.[locationName] &&
-              normalizeTimestampToMilliseconds(user.lastSeenIn[locationName]) >
-                lastSeenThreshold
-          )
-        : [],
-      isRecentLocationUsersLoaded: isWorldUsersLoaded,
-    }),
-    [worldUsers, isWorldUsersLoaded, lastSeenThreshold, locationName]
-  );
-};
+export { useRecentLocationUsers } from "./useRecentLocationUsers";
 
 export const useRecentVenueUsers = () => {
   const venueId = useVenueId();
